add message-box dialog

+++ src/app/modules/expenses/expense-list/expense-list.component.ts
@@ -7,6 +7,14 @@ import { Router } from '@angular/router';
 import { NotificationService } from '../../../shared/services/notification.service';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
+import {
+	MessageBox,
+	MessageIcon,
+	MessageBoxButton,
+	MessageResult,
+} from 'src/app/components/message-dialog/message-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
 	selector: 'app-expense-list',
@@ -37,9 +45,11 @@ export class ExpenseListComponent implements OnInit {
 	editField: string;
 
 	constructor(
+		private router: Router,
 		private expenseService: ExpenseService,
 		private notification: NotificationService,
-		private router: Router
+		private translateService: TranslateService,
+		private dialog: MatDialog
 	) {}
 
 	ngOnInit() {
@@ -64,29 +74,43 @@ export class ExpenseListComponent implements OnInit {
 		this.expenses[idx][property] = editField;
 	}
 
-	remove(idx: any, expense: IExpense) {
-		this.expenseService.deleteExpense(expense.id).subscribe(
-			data => {
-				console.log(data);
-				this.expenses.splice(idx, 1);
-				if (expense.imageId)
-					this.expenseService
-						.deleteExpenseImage(expense.imageId)
-						.subscribe(
-							data =>
-								this.notification.success(
-									'Expense has been deleted.',
-									'Delete an Expense'
-								),
-							error =>
-								this.notification.error(
-									`Expense has not been deleted. Error: ${error}`,
-									'Delete an Expense'
-								)
+	remove(expense: IExpense) {
+		MessageBox.show(
+			this.dialog,
+			this.translateService.instant('delete_message_box_question'),
+			this.translateService.instant('delete_message_box_tittle'),
+			MessageBoxButton.YesNo,
+			null,
+			false,
+			MessageIcon.Question
+		).subscribe(data => {
+			if (data.result === MessageResult.Yes) {
+				this.expenseService.deleteExpense(expense.id).subscribe(
+					data => {
+						console.log(data);
+						this.expenses = this.expenses.filter(
+							item => item.id == expense.id
 						);
-			},
-			error => console.log(error)
-		);
+						if (expense.imageId)
+							this.expenseService
+								.deleteExpenseImage(expense.imageId)
+								.subscribe(
+									data =>
+										this.notification.success(
+											'Expense has been deleted.',
+											'Delete an Expense'
+										),
+									error =>
+										this.notification.error(
+											`Expense has not been deleted. Error: ${error}`,
+											'Delete an Expense'
+										)
+								);
+					},
+					error => console.log(error)
+				);
+			}
+		});
 	}
 
 	add() {
